server/requirements.txt
psycopg2-binary
Django
python-dotenv
djangorestframework
djangorestframework-simplejwt
PyJWT
